+++ libraries/vmhost_ntp.rb
@@ -1,16 +1,16 @@
-require 'vsphere'
+require 'esx_conn'
 
 # Custom resource based on the InSpec resource DSL
-class VmWareHostLockdown < Inspec.resource(1)
-  name 'host_lockdown'
+class VmWareHostNtpServer < Inspec.resource(1)
+  name 'vmhost_ntpserver'
 
   desc "
-    This resources reads the actual build number of a hostsystem.
+    This resource reads the ntp configuration of a hostsystem.
   "
 
   example "
-    describe host_lockdown({datacenter: 'ha-datacenter', host: 'localhost') do
-      it { should be_enabled }
+    describe host_ntpserver({datacenter: 'ha-datacenter', host: 'localhost', ntp: '0.europe.pool.ntp.org'}) do
+      it { should exist }
     end
   "
 
@@ -24,18 +24,15 @@ class VmWareHostLockdown < Inspec.resource(1)
     [name.to_s]
   end
 
-  def enabled?
+  def exists?
     host = get_host(@opts[:datacenter], @opts[:host])
-    host.config.lockdownMode.include?('Enabled')
-  end
-
-  def disabled?
-    !enabled?
+    options = host.config.dateTimeInfo.ntpConfig.server
+    options.include?(@opts[:ntp])
   end
 
   def get_host(dc_name, host_name)
     # TODO: this should something like `inspec.vsphere.connection`
-    vim = VSphere.new.connection
+    vim = ESXConnection.new.connection
     dc = vim.serviceInstance.find_datacenter(dc_name)
     hosts = dc.hostFolder.children
     hosts.each do |entity|
