--- libraries/vmware_vm_advancedsetting.rb
@@ -1,51 +1,0 @@
-require 'vsphere'
-
-# Custom resource based on the InSpec resource DSL
-class VmWareConfig < Inspec.resource(1)
-  name 'vmware_vm_advancedsetting'
-
-  desc "
-    This resources reads all vm advanced configuration options.
-  "
-
-  example "
-    describe vmware_vm_advancedsetting({datacenter: 'ha-datacenter', vm: '1'}) do
-      its('softPowerOff') { should eq 'false' }
-    end
-  "
-
-  # Load the configuration file on initialization
-  def initialize(opts)
-    @opts = opts
-  end
-
-  # Expose all parameters
-  def method_missing(name) # rubocop:disable Style/MethodMissing
-    advancedsetting[name.to_s]
-  end
-
-  private
-
-  def advancedsetting
-    return @params if defined?(@params)
-    vm = get_vm(@opts[:datacenter], @opts[:vm])
-    if vm.nil?
-      @params = {}
-    else
-      # convert to key value pairs
-      @params = {}
-      vm.config.extraConfig.each { |item|
-        @params[item.key] = item.value
-      }
-      @params
-    end
-  end
-
-  def get_vm(datacenter_name, vm_name)
-    # TODO: this should something like `inspec.vsphere.connection`
-    vim = VSphere.new.connection
-    dc = vim.serviceInstance.find_datacenter(datacenter_name)
-    vm = dc.find_vm(vm_name)
-    vm
-  end
-end
